:art: refactored code
